Cleaned up a bit :)
