+++ src/App.js
@@ -1,12 +1,12 @@
 import './App.css';
-import Navbar from './components/Navbar';
+import Navbar from './components/NavBar/Navbar';
 import Info from './components/Info';
-import LinkTree from './components/LinkTree/LinkTree';
+import LinkTree from './components/MobileVersion/LinkTree/LinkTree';
 
 function App() {
   return (
       <div className="App">
-        <div class="visibledesktop">
+        <div className="visibledesktop">
           <header className="header">
             <h1>
               Taheera Ahmed
@@ -19,10 +19,10 @@ function App() {
           <footer className="footer">
           </footer>
         </div>
-        <div class="visibledevice">
+        <div className="visibledevice">
 
           <LinkTree/>
-          <div class="footerDevice">
+          <div className="footerDevice">
             Liten heads-up: Du ser på mobilversjonen av nettsiden. 
             Dersom du ønsker å se hele nettsiden må du bruke en enhet med større skjerm
             Cool bye :3
