+++ src/components/NavBar/Frontpage/Highlighter/Highlighter.js
