+++ src/components/NavBar/Navbar.js
@@ -1,0 +1,128 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { makeStyles } from '@material-ui/core/styles';
+import AppBar from '@material-ui/core/AppBar';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
+import { ProjectList } from './Projects/ProjectList.js'
+import {FrontPage} from "./Frontpage/Frontpage";
+import {CV} from "./CV/CV";
+
+
+function TabPanel(props) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`nav-tabpanel-${index}`}
+      aria-labelledby={`nav-tab-${index}`}
+      {...other}
+    >
+      {value === index && (
+        <Box p={3}>
+          <Typography>{children}</Typography>
+        </Box>
+      )}
+    </div>
+  );
+}
+
+TabPanel.propTypes = {
+  children: PropTypes.node,
+  index: PropTypes.any.isRequired,
+  value: PropTypes.any.isRequired,
+};
+
+function a11yProps(index) {
+  return {
+    id: `nav-tab-${index}`,
+    'aria-controls': `nav-tabpanel-${index}`,
+  };
+}
+
+function LinkTab(props) {
+  return (
+    <Tab
+      component="a"
+      onClick={(event) => {
+        event.preventDefault();
+      }}
+      {...props}
+    />
+  );
+}
+
+const useStyles = makeStyles({
+  root: {
+    background: 'white',
+  },
+  indicator: {
+    border: 'double 2px transparent',
+    backgroundImage: 'linear-gradient(transparent, transparent), radial-gradient(circle at top left, #FF8E53 30%,#FE6B8B 90% )',
+  },
+  tabs: {
+    background: "linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)",
+    fontSize: 20,
+  },
+  h3: {
+    textAlign: 'left',
+    fontSize: 25,
+    paddingTop:30,
+  }
+});
+
+
+
+export default function Navbar() {
+  const classes = useStyles();
+  const [value, setValue] = React.useState(0);
+
+  const handleChange = (event, newValue) => {
+    setValue(newValue);
+  };
+
+  return (
+    <div className={classes.root}>
+      <AppBar position="static" elevation={0}>
+        <Tabs
+          className={classes.tabs}
+          classes={{ indicator: classes.indicator }}
+          variant="fullWidth"
+          value={value}
+          onChange={handleChange}
+          aria-label="nav tabs example"
+          elevation={0}
+        >
+          <LinkTab label="Om meg" href="/drafts" {...a11yProps(0)} />
+          <LinkTab label="CV" href="/trash" {...a11yProps(1)} />
+          <LinkTab label="Prosjekter" href="/spam" {...a11yProps(2)} />
+          <LinkTab label="Fordi jeg kan" href="/spam" {...a11yProps(3)} />
+        </Tabs>
+      </AppBar>
+      <TabPanel value={value} index={0}>
+        <FrontPage/>
+      </TabPanel>
+      <TabPanel value={value} index={1}>
+        <CV/>
+      </TabPanel>
+      <TabPanel value={value} index={2}>
+        <Typography variant="h2" >
+            Prosjekter
+        </Typography>
+          <ProjectList/>
+      </TabPanel>
+      <TabPanel value={value} index={3}>
+        <Typography variant="h2" paragraph = {true}>
+          Fordi jeg kan
+        </Typography>
+        <Typography>
+          Her vil jeg legge inn morsomme ting 
+        </Typography>
+      </TabPanel>
+    </div>
+  );
+}
