+++ src/components/NavBar/Projects/Entries.js
