+++ src/components/NavBar/Projects/EntryProject.js
