+++ src/views/Dashboard/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Dashboard.css';
-import DigitalClock from '../../components/DigitalClock/DigitalClock';
-import FavoriteWebSites from '../../components/FavoriteWebSites/FavoriteWebSites';
+import DigitalClock from '../../components/Dashboard/DigitalClock/DigitalClock';
+import FavoriteWebSites from '../../components/Dashboard/FavoriteWebSites/FavoriteWebSites';
 
 
 export default class Dashboard extends React.Component{
